requirements.txt
bcrypt==5.0.0
blinker==1.9.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
Flask==3.1.2
Flask-Bcrypt==1.0.1
Flask-MySQLdb==2.0.0
Flask-SQLAlchemy==3.1.1
greenlet==3.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
mysqlclient==2.2.7
pillow==12.0.0
psycopg2-binary==2.9.11
PyMySQL==1.1.2
python-dotenv==1.2.1
reportlab==4.4.5
SQLAlchemy==2.0.44
typing_extensions==4.15.0
Werkzeug==3.1.4
